Fix windows warnings: macro collision of curses.h with wincon.h, 64 bit int assignement to 32 bit int for upate length, remove pop/push hack in pch.h since everything compiles fine under windows without it.

+++ src/TUIMenuWindow.cpp
@@ -121,7 +121,7 @@ void TUIMenuWindow::Run()
 
 void TUIMenuWindow::Update()
 {
-    for (int i = 0; i < mCommands.size(); ++i)
+    for (size_t i = 0; i < mCommands.size(); ++i)
     {
         wattrset(mWin, i == mOption ? A_REVERSE : A_NORMAL);
         mvwaddstr(mWin, 1 + i, 1, mCommands[i].first.c_str());
